banchat/requirements.txt
# requirements.txt
streamlit
requests
python-dotenv
google-generativeai